requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.6
black==21.6b0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-appconf==1.0.4
django-debug-toolbar==3.2.1
django-imagekit==4.0.2
django-phonenumber-field==5.2.0
django-redis-cache==3.0.0
django-taggit==1.4.0
django-uuslug==1.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.8.1
phonenumbers==8.12.24
pilkit==2.0
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
pylint==2.8.3
python-dotenv==0.17.1
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
regex==2021.7.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.5
wrapt==1.12.1
